+++ supertokens_python/recipe/emailpassword/api/implementation.py
@@ -16,14 +16,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from supertokens_python.asyncio import get_user
-from supertokens_python.auth_utils import (
-    SignInNotAllowedResponse,
-    SignUpNotAllowedResponse,
-    get_authenticating_user_and_add_to_current_tenant_if_required,
-    is_fake_email,
-    post_auth_checks,
-    pre_auth_checks,
-)
 from supertokens_python.logger import log_debug_message
 from supertokens_python.recipe.accountlinking import (
     AccountInfoWithRecipeIdAndUserId,
@@ -42,7 +34,6 @@ from supertokens_python.recipe.emailpassword.interfaces import (
     EmailExistsGetOkResult,
     GeneratePasswordResetTokenPostNotAllowedResponse,
     GeneratePasswordResetTokenPostOkResult,
-    LinkingToSessionUserFailedError,
     PasswordPolicyViolationError,
     PasswordResetPostOkResult,
     PasswordResetTokenInvalidError,
@@ -63,14 +54,24 @@ from supertokens_python.recipe.emailpassword.types import (
 from supertokens_python.recipe.emailverification.recipe import EmailVerificationRecipe
 from supertokens_python.recipe.session import SessionContainer
 from supertokens_python.recipe.totp.types import UnknownUserIdError
+from supertokens_python.types.auth_utils import LinkingToSessionUserFailedError
+from supertokens_python.types.base import AccountInfoInput
 from supertokens_python.types.response import GeneralErrorResponse
 
 from ..utils import get_password_reset_link
 
 if TYPE_CHECKING:
+    from supertokens_python.auth_utils import (
+        SignInNotAllowedResponse,
+        SignUpNotAllowedResponse,
+        get_authenticating_user_and_add_to_current_tenant_if_required,
+        is_fake_email,
+        post_auth_checks,
+        pre_auth_checks,
+    )
     from supertokens_python.recipe.emailpassword.interfaces import APIOptions
 
-from supertokens_python.types import AccountInfo, RecipeUserId
+from supertokens_python.types import RecipeUserId
 
 
 class APIImplementation(APIInterface):
@@ -84,7 +85,7 @@ class APIImplementation(APIInterface):
         # Check if there exists an email password user with the same email
         users = await AccountLinkingRecipe.get_instance().recipe_implementation.list_users_by_account_info(
             tenant_id=tenant_id,
-            account_info=AccountInfo(email=email),
+            account_info=AccountInfoInput(email=email),
             do_union_of_account_info=False,
             user_context=user_context,
         )
@@ -163,7 +164,7 @@ class APIImplementation(APIInterface):
 
         users = await AccountLinkingRecipe.get_instance().recipe_implementation.list_users_by_account_info(
             tenant_id=tenant_id,
-            account_info=AccountInfo(email=email),
+            account_info=AccountInfoInput(email=email),
             do_union_of_account_info=False,
             user_context=user_context,
         )
@@ -702,7 +703,7 @@ class APIImplementation(APIInterface):
         if pre_auth_check_res.status == "SIGN_UP_NOT_ALLOWED":
             conflicting_users = await AccountLinkingRecipe.get_instance().recipe_implementation.list_users_by_account_info(
                 tenant_id=tenant_id,
-                account_info=AccountInfo(
+                account_info=AccountInfoInput(
                     email=email,
                 ),
                 do_union_of_account_info=False,
