+++ supertokens_python/recipe/emailpassword/api/signin.py
@@ -15,10 +15,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Dict
 
-from supertokens_python.auth_utils import load_session_in_auth_api_if_needed
 from supertokens_python.recipe.emailpassword.interfaces import SignInPostOkResult
 
 if TYPE_CHECKING:
+    from supertokens_python.auth_utils import load_session_in_auth_api_if_needed
     from supertokens_python.recipe.emailpassword.interfaces import (
         APIInterface,
         APIOptions,
