+++ supertokens_python/recipe/emailpassword/asyncio/__init__.py
@@ -17,7 +17,6 @@ from typing_extensions import Literal
 
 from supertokens_python import get_request_from_user_context
 from supertokens_python.asyncio import get_user
-from supertokens_python.auth_utils import LinkingToSessionUserFailedError
 from supertokens_python.recipe.emailpassword import EmailPasswordRecipe
 from supertokens_python.recipe.emailpassword.interfaces import (
     ConsumePasswordResetTokenOkResult,
@@ -38,6 +37,7 @@ from supertokens_python.recipe.emailpassword.types import (
 )
 from supertokens_python.recipe.emailpassword.utils import get_password_reset_link
 from supertokens_python.recipe.session import SessionContainer
+from supertokens_python.types.auth_utils import LinkingToSessionUserFailedError
 
 from ....types import RecipeUserId
 from ...multitenancy.constants import DEFAULT_TENANT_ID
