+++ supertokens_python/recipe/emailpassword/recipe_implementation.py
@@ -16,15 +16,12 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict, Union
 
 from supertokens_python.asyncio import get_user
-from supertokens_python.auth_utils import (
-    LinkingToSessionUserFailedError,
-    link_to_session_if_provided_else_create_primary_user_id_or_link_by_account_info,
-)
 from supertokens_python.normalised_url_path import NormalisedURLPath
 from supertokens_python.recipe.accountlinking.recipe import AccountLinkingRecipe
 from supertokens_python.recipe.emailverification.recipe import EmailVerificationRecipe
 from supertokens_python.recipe.session import SessionContainer
 from supertokens_python.types import RecipeUserId
+from supertokens_python.types.auth_utils import LinkingToSessionUserFailedError
 
 from ...types import User
 from .constants import FORM_FIELD_PASSWORD_ID
@@ -45,6 +42,9 @@ from .interfaces import (
 from .utils import EmailPasswordConfig
 
 if TYPE_CHECKING:
+    from supertokens_python.auth_utils import (
+        link_to_session_if_provided_else_create_primary_user_id_or_link_by_account_info,
+    )
     from supertokens_python.querier import Querier
