+++ supertokens_python/recipe/passwordless/interfaces.py
@@ -18,7 +18,6 @@ from typing import Any, Dict, List, Optional, Union
 
 from typing_extensions import Literal
 
-from supertokens_python.auth_utils import LinkingToSessionUserFailedError
 from supertokens_python.framework import BaseRequest, BaseResponse
 from supertokens_python.ingredients.emaildelivery import EmailDeliveryIngredient
 from supertokens_python.recipe.session import SessionContainer
@@ -26,6 +25,7 @@ from supertokens_python.types import (
     RecipeUserId,
     User,
 )
+from supertokens_python.types.auth_utils import LinkingToSessionUserFailedError
 from supertokens_python.types.response import APIResponse, GeneralErrorResponse
 
 from ...supertokens import AppInfo
