+++ supertokens_python/recipe/passwordless/recipe.py
@@ -18,7 +18,6 @@ from typing import TYPE_CHECKING, Any, Awaitable, Callable, Dict, List, Optional
 
 from typing_extensions import Literal
 
-from supertokens_python.auth_utils import is_fake_email
 from supertokens_python.ingredients.emaildelivery import EmailDeliveryIngredient
 from supertokens_python.ingredients.emaildelivery.types import EmailDeliveryConfig
 from supertokens_python.ingredients.smsdelivery import SMSDeliveryIngredient
@@ -78,6 +77,7 @@ from .utils import (
 )
 
 if TYPE_CHECKING:
+    from supertokens_python.auth_utils import is_fake_email
     from supertokens_python.framework.request import BaseRequest
     from supertokens_python.framework.response import BaseResponse
     from supertokens_python.supertokens import AppInfo
