+++ supertokens_python/recipe/webauthn/interfaces/recipe.py
@@ -11,10 +11,10 @@ from typing import (
 
 from typing_extensions import NotRequired, Unpack
 
-from supertokens_python.auth_utils import LinkingToSessionUserFailedError
 from supertokens_python.recipe.session.interfaces import SessionContainer
 from supertokens_python.recipe.webauthn.types.base import UserContext
 from supertokens_python.types import RecipeUserId, User
+from supertokens_python.types.auth_utils import LinkingToSessionUserFailedError
 from supertokens_python.types.response import (
     CamelCaseBaseModel,
     OkResponseBaseModel,
