+++ supertokens_python/types/auth_utils.py
@@ -1,0 +1,25 @@
+from typing_extensions import Literal
+
+from supertokens_python.types.response import StatusReasonResponseBaseModel
+
+
+class LinkingToSessionUserFailedError(
+    StatusReasonResponseBaseModel[
+        Literal["LINKING_TO_SESSION_USER_FAILED"],
+        Literal[
+            "EMAIL_VERIFICATION_REQUIRED",
+            "RECIPE_USER_ID_ALREADY_LINKED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+            "ACCOUNT_INFO_ALREADY_ASSOCIATED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+            "SESSION_USER_ACCOUNT_INFO_ALREADY_ASSOCIATED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+            "INPUT_USER_IS_NOT_A_PRIMARY_USER",
+        ],
+    ]
+):
+    status: Literal["LINKING_TO_SESSION_USER_FAILED"] = "LINKING_TO_SESSION_USER_FAILED"
+    reason: Literal[
+        "EMAIL_VERIFICATION_REQUIRED",
+        "RECIPE_USER_ID_ALREADY_LINKED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+        "ACCOUNT_INFO_ALREADY_ASSOCIATED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+        "SESSION_USER_ACCOUNT_INFO_ALREADY_ASSOCIATED_WITH_ANOTHER_PRIMARY_USER_ID_ERROR",
+        "INPUT_USER_IS_NOT_A_PRIMARY_USER",
+    ]
